Update 'more-vscode' patches for Isabelle repo, at 2021-03-25.
This patch brings us from Isabelle2021 (hg*7e2a9a8c2b85) to the current
(as of now) repository revision, hg*6b480efe1bc3.  It incorporates:

 + hg*0ffcad1f6130 (wenzelm):  tuned --- fewer warnings;
 + hg*31d4274f32de (wenzelm):  tuned --- fewer warnings;
 + hg*d8a0e996614b (wenzelm):  tuned --- fewer warnings;
 + hg*77ef8bef0593 (wenzelm):  clarified signature --- fewer warnings;

(I use Github's Subversion frontend to replace only src/Tools/VSCode
in a repository version of Isabelle with this.  It seems to work.)

I have only a tyro's grasp of Scala or of the guts of Isabelle.
This commit is in no way an admission that I know what I'm doing.

[mq-changeset: b87b62f7420c isa2021-more-vscode-upd-20210325 r18]